+++ CHANGELOG.md
@@ -55,6 +55,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   > Although the tsconfig.json has `"noUnusedParameters": false`,  
   > adding the underscore prevent warnings in case of is set to `true`,  
   > and also helps to understand better unused parameters for future cleanups.  
+- Add support for array parameters to rgb functions in `color-functions` plugin.
 
 
 ## [1.4.0] -